Separate functions for reading and writing within the base command.
